# app/static/swagger.yaml
swagger: "2.0"
info:
  title: "Mechanic Shop API"
  description: "API documentation for the mechanic shop management system"
  version: "1.0.0"
host: "127.0.0.1:5000"
schemes:
  - "http"
consumes:
  - "application/json"
produces:
  - "application/json"
securityDefinitions:
  bearerAuth:
    type: apiKey
    name: Authorization
    in: header
...
paths:
...
  /users/register:
    post:
      tags: [Users]
      summary: "Register a new user"
      description: "Creates a new user account."
      parameters:
        - in: body
          name: user
          required: true
          schema:
            $ref: "#/definitions/UserRegisterPayload"
      responses:
        201:
          description: "User registered"
          schema:
            $ref: "#/definitions/User"
        400:
          description: "Missing fields or email exists"
  /users/login:
    post:
      tags: [Users]
      summary: "Login user"
      description: "Authenticates a user and returns a JWT token."
      parameters:
        - in: body
          name: credentials
          required: true
          schema:
            $ref: "#/definitions/UserLoginPayload"
      responses:
        200:
          description: "Login successful"
          schema:
            $ref: "#/definitions/UserLoginResponse"
        400:
          description: "Validation error"
        401:
          description: "Invalid credentials"
  /users:
    get:
      tags: [Users]
      summary: "Get all users"
      description: "Admin only."
      security:
        - bearerAuth: []
      responses:
        200:
          description: "List of users"
          schema:
            type: array
            items:
              $ref: "#/definitions/User"
  /users/{user_id}:
    get:
      tags: [Users]
      summary: "Get a single user"
      description: "Admins and mechanics can retrieve a user."
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: user_id
          required: true
          type: integer
      responses:
        200:
          description: "User found"
          schema:
            $ref: "#/definitions/User"
        404:
          description: "User not found"

    put:
      tags: [Users]
      summary: "Update a user"
      description: "Admin only."
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: user_id
          required: true
          type: integer
        - in: body
          name: user
          required: true
          schema:
            $ref: "#/definitions/UserUpdatePayload"
      responses:
        200:
          description: "User updated"
          schema:
            $ref: "#/definitions/UserUpdateResponse"
        400:
          description: "Invalid role or email exists"
        404:
          description: "User not found"

    delete:
      tags: [Users]
      summary: "Delete a user"
      description: "Admin only."
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: user_id
          required: true
          type: integer
      responses:
        200:
          description: "User deleted"
          schema:
            $ref: "#/definitions/UserDeleteResponse"
        404:
          description: "User not found"

  # MECHANICS
...
  /vehicles:
    get:
      tags: [Vehicles]
      summary: "Get all vehicles"
      description: "Returns all vehicles."
      responses:
        200:
          description: "List of vehicles"
          schema:
            type: array
            items:
              $ref: "#/definitions/Vehicle"

    post:
      tags: [Vehicles]
      summary: "Create a vehicle"
      description: "Admins and mechanics only."
      security:
        - bearerAuth: []
      parameters:
        - in: body
          name: vehicle
          required: true
          schema:
            $ref: "#/definitions/VehicleCreatePayload"
      responses:
        201:
          description: "Vehicle created"
          schema:
            $ref: "#/definitions/VehicleCreateResponse"
        400:
          description: "Validation error"
        409:
          description: "VIN exists"
  /vehicles/{vehicle_id}:
    get:
      tags: [Vehicles]
      summary: "Get a vehicle"
      description: "Retrieve a vehicle by ID."
      parameters:
        - in: path
          name: vehicle_id
          required: true
          type: integer
      responses:
        200:
          description: "Vehicle found"
          schema:
            $ref: "#/definitions/Vehicle"
        404:
          description: "Vehicle not found"

    put:
      tags: [Vehicles]
      summary: "Update a vehicle"
      description: "Admins and mechanics only."
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: vehicle_id
          required: true
          type: integer
        - in: body
          name: vehicle
          required: true
          schema:
            $ref: "#/definitions/VehicleUpdatePayload"
      responses:
        200:
          description: "Vehicle updated"
          schema:
            $ref: "#/definitions/VehicleUpdateResponse"
        400:
          description: "Validation error"
        404:
          description: "Vehicle not found"
        409:
          description: "VIN exists"

    delete:
      tags: [Vehicles]
      summary: "Delete a vehicle"
      description: "Admin only."
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: vehicle_id
          required: true
          type: integer
      responses:
        200:
          description: "Vehicle deleted"
          schema:
            $ref: "#/definitions/VehicleDeleteResponse"
        404:
          description: "Vehicle not found"

  # SERVICE TICKETS
...
definitions:
...
  User:
    type: object
    properties:
      id:
        type: integer
      name:
        type: string
      email:
        type: string
      role:
        type: string
        enum: [admin, mechanic, customer]
      created_at:
        type: string
      updated_at:
        type: string
  UserRegisterPayload:
    type: object
    required: [name, email, password]
    properties:
      name:
        type: string
      email:
        type: string
      password:
        type: string
  UserLoginPayload:
    type: object
    required: [email, password]
    properties:
      email:
        type: string
      password:
        type: string
  UserLoginResponse:
    type: object
    properties:
      token:
        type: string
      user:
        $ref: "#/definitions/User"
  UserUpdatePayload:
    type: object
    properties:
      name:
        type: string
      email:
        type: string
      role:
        type: string
        enum: [admin, mechanic, customer]
  UserUpdateResponse:
    type: object
    properties:
      message:
        type: string
      user:
        $ref: "#/definitions/User"
  UserDeleteResponse:
    type: object
    properties:
      message:
        type: string

  # MECHANICS
...
  Vehicle:
    type: object
    properties:
      id:
        type: integer
      customer_id:
        type: integer
      make:
        type: string
      model:
        type: string
      year:
        type: integer
      vin:
        type: string
  VehicleCreatePayload:
    type: object
    required: [customer_id, make, model, year, vin]
    properties:
      customer_id:
        type: integer
      make:
        type: string
      model:
        type: string
      year:
        type: integer
      vin:
        type: string
  VehicleCreateResponse:
    type: object
    properties:
      message:
        type: string
      vehicle:
        $ref: "#/definitions/Vehicle"
  VehicleUpdatePayload:
    type: object
    properties:
      make:
        type: string
      model:
        type: string
      year:
        type: integer
      vin:
        type: string
  VehicleUpdateResponse:
    type: object
    properties:
      message:
        type: string
      vehicle:
        $ref: "#/definitions/Vehicle"
  VehicleDeleteResponse:
    type: object
    properties:
      message:
        type: string

  # SERVICE TICKETS
...